third_party/blink/web_tests/external/wpt/fetch/private-network-access/redirect.https.window-expected.txt

This is a testharness.js-based test.
PASS local to private to public: success.
PASS local to private to local: success.
PASS private to private to local: failed preflight.
PASS private to private to local: success.
PASS private to private to local: no-cors success.
PASS private to local to private: failed preflight.
FAIL private to local to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
FAIL private to local to private: no-cors success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
PASS public to private to local: failed first preflight.
PASS public to private to local: failed second preflight.
FAIL public to private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
FAIL public to private to local: no-cors success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
Harness: the test ran to completion.
